Continued work on ringzero: * Generalised de::PathDirectory so it can be used for modelling anything as a hierarchical string-key'd tree. Renamed the original to de::FileDirectory * When a cvar callback returns a pointer to the template used to create it it should be const. Do we even want to do this? I would think some other kind of handle would be better (we can then remove the pretty much unused copy in ddcvar_t). * When destroying the console databases get rid of the knownWords first, as this prevents the other database destructors from doing additional redundant work. * Clean up.

+++ doomsday/engine/portable/src/dam_file.c
@@ -1158,7 +1158,7 @@ static void archiveMaterialDict(materialdict_t *dict, boolean write)
         dict->count = readLong();
         for(i = 0; i < dict->count; ++i)
         {
-            size_t len = readLong();
+            int len = readLong();
             Str_Clear(&dict->table[i].path);
             Str_Reserve(&dict->table[i].path, len);
             readNBytes(Str_Text(&dict->table[i].path), len);
